+++ packages/create-bembajs/templates/base/vite.config.mjs
@@ -1,0 +1,12 @@
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+import { vitePluginBemba } from 'bembajs-core/vite-plugin-bemba.js';
+
+export default defineConfig({
+  plugins: [vitePluginBemba(), react()],
+  resolve: {
+    extensions: ['.bemba', '.jsx', '.js', '.tsx', '.ts', '.json']
+  },
+  server: { port: 3000 },
+  build: { outDir: 'dist' }
+});